+++ database/starters/oracle-spring-boot-starter-samples/oracle-spring-boot-sample-ucp-jpa/src/test/java/com/oracle/database/spring/sample/UCPSampleApplicationTest.java
@@ -1,9 +1,11 @@
-// Copyright (c) 2024, Oracle and/or its affiliates.
+// Copyright (c) 2024, 2026, Oracle and/or its affiliates.
 // Licensed under the Universal Permissive License v 1.0 as shown at https://oss.oracle.com/licenses/upl.
 package com.oracle.database.spring.sample;
 
 import java.time.Duration;
 import java.util.List;
+
+import javax.sql.DataSource;
 
 import io.micrometer.core.instrument.FunctionCounter;
 import io.micrometer.core.instrument.Gauge;
@@ -52,8 +54,11 @@ public class UCPSampleApplicationTest {
     @Autowired
     MeterRegistry meterRegistry;
 
+    @Autowired
+    DataSource dataSource;
+
     @Test
-    void ucpSampleApp() {
+    void ucpSampleApp() throws Exception {
 
         // Create a new student
         Student student1 = new Student(
@@ -82,8 +87,13 @@ public class UCPSampleApplicationTest {
         ResponseEntity<Student> re = studentController.getStudent(s1.getId());
         assertThat(re.getStatusCode()).isEqualTo(HttpStatusCode.valueOf(404));
 
-        Gauge activeConnections = meterRegistry.get("ucp.connections.active")
-                .tag("pool", "UCPSampleApplication")
+        Gauge idleConnections = meterRegistry.get("db.client.connection.count")
+                .tag("db.client.connection.pool.name", "UCPSampleApplication")
+                .tag("db.client.connection.state", "idle")
+                .gauge();
+        Gauge usedConnections = meterRegistry.get("db.client.connection.count")
+                .tag("db.client.connection.pool.name", "UCPSampleApplication")
+                .tag("db.client.connection.state", "used")
                 .gauge();
         FunctionCounter borrowedConnections = meterRegistry.get("ucp.connections.borrowed")
                 .tag("pool", "UCPSampleApplication")
@@ -91,11 +101,14 @@ public class UCPSampleApplicationTest {
         FunctionCounter returnedConnections = meterRegistry.get("ucp.connections.returned")
                 .tag("pool", "UCPSampleApplication")
                 .functionCounter();
-        Gauge maxConnections = meterRegistry.get("ucp.connections.max")
-                .tag("pool", "UCPSampleApplication")
+        Gauge maxConnections = meterRegistry.get("db.client.connection.max")
+                .tag("db.client.connection.pool.name", "UCPSampleApplication")
                 .gauge();
 
-        assertThat(activeConnections).isNotNull();
+        assertThat(idleConnections.value()).isGreaterThan(0);
+        try (var ignored = dataSource.getConnection()) {
+            assertThat(usedConnections.value()).isGreaterThan(0);
+        }
         assertThat(borrowedConnections.count()).isGreaterThan(0);
         assertThat(returnedConnections.count()).isGreaterThan(0);
         assertThat(maxConnections.value()).isEqualTo(30.0);